moved changeStyleAtribut to own class
